Add JPA for [member]

+++ src/main/java/vn/elca/training/dao/IGroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import vn.elca.training.dom.Group;
 
-public interface IGroupRepository extends CrudRepository<Group, Integer>{
+public interface IGroupRepository extends CrudRepository<Group, Long>{
 
 }
